+++ FrbaHotel/cliente.cs
@@ -18,7 +18,7 @@ namespace FrbaHotel
         public string nombre { get; set; }
         public DateTime? fecha_nac { get; set; }
         public string mail { get; set; }
-        public int? tel { get; set; }
+        public long? tel { get; set; }
         public string calle { get; set; }
         public int? nro_calle { get; set; }
         public int? piso { get; set; }
